When a Screen Transition begins, the Camera will start moving. When the Camera reaches the end Transition Point, it will stop moving.
