--- RWM-P2-TEAM-C/Assets/Scripts/ScreenTransitions.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 55823df5213318741bf3d222a3f6f827
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
