--- RWM-P2-TEAM-C/Assets/Scripts/ScreenTransitions/CameraMover.cs
@@ -1,18 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CameraMover : MonoBehaviour
-{
-    public Vector2 transitionPoint;
-
-    void StartMovement()
-    {
-        transitionPoint = this.gameObject.GetComponent<ScreenTransition>().transitionPoint;
-    }
-
-    IEnumerator movement()
-    {
-
-    }
-}
